abi.encodePacked -> abi.encode
use the standardized abi.encode for workload id calculation, as this is standardized
and will ease the process of decoding it back to its constituent data for both
onchain and offchain users.

See https://docs.soliditylang.org/en/latest/abi-spec.html\#non-standard-packed-mode
for more detailed info on the difference between the two functions.