make MockAutomataDcapAttestationFee handle multiple quote inputs
We are blocked on deploying to experimental because we cannot get the
Automata contracts all deployed and loaded with proper collateral values.
So in order to unblock us we are going to deploy a Mock of the Automata
contract Flashtestations uses to interact with the Automata verification
process (test/mocks/MockAutomataDcapAttestationFee.sol).