add commitHash and recordLocators to workload when adding workloads to a policy
Completes UNI-835

We want to add the commit hash to the Policy because Flashtestations
needs a way to associate a TEE's workloadId with the source code used
to build the TEE image represented by that workloadId. There is no way
to do this that does not involve some sort of onchain permissioned action,
because to do it without permissioned action would require building TEE
images onchain, which is prohibitively expensive. Instead, we settle for
a multisig signer that is permissioned to associate workloadIds with commit hashes.

The recordLocators part of the metadata makes it easier for verifiers to find the URL from
which to fetch the source code. We provide multiple locators so there is higher availability.

Offchain verifiers can then use this commit hash to locate the TEE image build source code,
build the TEE image, derive the workloadId, and then compare that locally-built
workloadId with the workloadId in the policy that is associated with the commit hash.
In this way, anyone can reliably prove that a given block is built using source code
that orders transactions in a fair and verifiable manner which is the purpose of flashtestations).

fix broken tests

fixed tests