remove supported version checks on _verifyBlockBuilderProof
This addresses the Q3 2025 OZ audit item M-02, which pointed out that there
are no onchain checks that verify the proof was built according to
the version passed in the '_verifyBlockBuilderProof' function call.

This is correct, there are no checks, and in fact there is no way to perform
these checks since the 'blockContentHash' portion of the proof is generated
entirely offchain. So rather than perform these needless 'isSupportedVersion'
checks, we are going to remove that check to save gas, and we assume that the
version passed in the function call is not to be trusted.

Instead, it serves as a notice to any offchain verifiers reading the event data
that this is the version of the flashtestation proof they should assume when
verifying the proof. It could be the wrong version, but that's fine because the
proving code is assumed to perform adequate checks if the version is wrong