L-04 add missing reentrancy guard in FlashtestationRegistry:invalidateAttestation
This addresses the L-04 item of the Q3 2025 OZ audit

Functions that register new TEE services have a reentrancy guard modifier because they do
not follow the CEI pattern and rely on the external call verifyAndAttestOnChain from the
attestation contract. While this call can be trusted because the attestation contract should not
be malicious and is not upgradeable, there are situations where the execution is forwarded to
the tx.origin.

Hence, with the recent introduction of EIP-7702, it is possible for a malicious user to reenter
the FlashtestationRegistry with an outdated state. For this reason, the reentrancy
guard is completely necessary. However, while the invalidateAttestation function relies
on this external call, it does not have the reentrancy guard protection. Hence, it is vulnerable to being reentered.

Consider adding the nonReentrant modifier to the invalidateAttestation function.