Fix L-05 no checks on sourceLocators array
This fixes L-05 on the Q3 2025 OZ audit.

When a new workload is added to a policy, the governance has to provide the commitHash
and sourceLocators to save the metadata for the workload. The commitHash is the
commit hash of the git repository whose source code is used to build the TEE image identified
by the workloadId, and the sourceLocators is an array of URIs pointing to that source
code. The commitHash cannot be empty because the length must be greater than 0, but
sourceLocators has no checks. This array should have at least one element to fetch the
source code.

Consider ensuring that the sourceLocators array has at least one element.