L-02 add missing deadline protection on FlashtestationRegistry:permitRegisterTEEService
This addresses L-02 of the Q3 2025 OZ audit:

When constraining permit signatures, it is highly recommended to have a deadline protection in
case a signature is not executed within a specific time window set by the signer.
Throughout the codebase, multiple instances of missing deadline protection for signatures
were identified:

• In FlashtestationRegistry.sol, the permitRegisterTEEService function
• In BlockBuilderPolicy.sol, the permitVerifyBlockBuilderProof function

Consider checking that the provided deadline is not expired and is part of the EIP-712 digest
hash.

Note: we do not add a deadline for BlockBuilderPolicy's permitVerifyBlockBuilderProof, because
the high frequency with which we'll be broadcasting new calls to permitVerifyBlockBuilderProof which invalidate the old calls makes a deadline not very useful for it. We intend to broadcast 1 such call every 200ms, for each flashblock. In fact, we can't even correctly implement a deadline for permitVerifyBlockBuilderProof, because the smallest granularity we could use for the deadline is 1 second, and that is not enough time to specify multiple flashblocks (which we'll need to do).