L-03 added invalidatePreviousSignature function
This addresses L-03 of the Q3 2025 OZ audit:

When a user signs a permit digest with a specific nonce but later wants to invalidate previous
signatures, the most used implementations are either a function to increment the nonce or
executing a no-op operation. In the case of FlashtestationRegistry, neither of these
two options is available. No-op permit executions are reverted in order to avoid gas costs.
Thus, the only way to increment the nonce is by executing a valid signature with a different
valid rawQuote value.

Consider adding a function to increment the nonce of the caller in order to invalidate previous
signatures not yet executed.