N-01 incomplete docstring
This completes N-01 of the Q3 2025 OZ audit:

In the initialize function of the FlashtestationRegistry contract, the owner parameter is not documented.

Consider thoroughly documenting all functions/events (and their parameters or return values) that are part of a contract's public API. When writing docstrings, consider following the Ethereum Natural Specification Format (NatSpec).