add indexes to IFlashtestationRegistry and IBlockBuilderPolicy events
This addresses N-05 of the Q3 2025 OZ audit:

Throughout the codebase, several events do not have indexed parameters:

The WorkloadAddedToPolicy event of BlockBuilderPolicy.sol.

The WorkloadRemovedFromPolicy event of BlockBuilderPolicy.sol.

The RegistrySet event of BlockBuilderPolicy.sol.

The BlockBuilderProofVerified event of BlockBuilderPolicy.sol.

The TEEServiceRegistered event of IFlashtestationRegistry.sol.

The TEEServiceInvalidated event of IFlashtestationRegistry.sol.

To improve the ability of off-chain services to search and filter for specific events, consider indexing event parameters.

Note: we didn't add it to the BlockBuilderProofVerified event, in order to keep that event as efficient as possible